+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## 0.1.6
+
+* bugfix export script
+
+## 0.1.5
+
+* barman export script
+
 ## 0.1.3
 
 * bugfix barman config file